requirements.txt
Flask==3.0.2
pymongo==4.6.1
flask-login==0.6.3
python-dotenv==1.0.1
APScheduler==3.10.4
bcrypt==4.1.2
Flask-PyMongo==2.3.0
